Reorganise code
- move component auditing into a separate file, like the application auditing and comparing code
- call everything from the controller
