+++ app/views/govuk_publishing_components/audit/show.html.erb
@@ -134,7 +134,7 @@
               </tr>
             </thead>
             <tbody class="govuk-table__body">
-              <% @gem[:component_listing].each do |component| %>
+              <% @components[:component_listing].each do |component| %>
                 <tr class="govuk-table__row">
                   <th scope="row" class="govuk-table__header">
                     <a href="<%= component[:link] %>" class="govuk-link"><%= component[:name] %></a>
@@ -184,7 +184,7 @@
         </div>
         <div id="accordion-default-content-2" class="govuk-accordion__section-content" aria-labelledby="accordion-default-heading-2">
           <dl class="govuk-summary-list">
-            <% @gem[:components_containing_components].each do |component| %>
+            <% @components[:components_containing_components].each do |component| %>
               <div class="govuk-summary-list__row">
                 <dt class="govuk-summary-list__key">
                   <a href="<%= component[:link] %>" class="govuk-link"><%= component[:component] %></a>
@@ -210,7 +210,7 @@
         </div>
         <div id="accordion-default-content-2" class="govuk-accordion__section-content" aria-labelledby="accordion-default-heading-2">
           <dl class="govuk-summary-list">
-            <% @gem[:components_by_application].each do |component| %>
+            <% @components[:components_by_application].each do |component| %>
               <div class="govuk-summary-list__row">
                 <dt class="govuk-summary-list__key">
                   <%= component[:component] %> (<%= component[:count] %>)
